Adds Reservation request manager
